refactor: organize registration components into dedicated subfolders
