+++ src/components/registration/shared/PasswordStrengthMeter.tsx
